# docs/src/external/_static/proposal_only_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
    "/ska-oso-services/oso/api/v2/pht/prsls/osd/{cycle}": {
      "get": {
        "tags": [
          "PPT API - Proposal Preparation"
        ],
        "summary": "Retrieve OSD data for a given cycle",
        "description": "This queries the OSD data by cycle id. \n\nThis data is made available for the PHT UI.\n\nReturns:\n    OsdDataModel: The OSD data validated against the defined schema.",
        "operationId": "get_osd_by_cycle_ska_oso_services_oso_api_v2_pht_prsls_osd__cycle__get",
        "parameters": [
          {
            "name": "cycle",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Cycle"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OsdDataModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ska-oso-services/oso/api/v2/pht/prsls/create": {
      "post": {
        "tags": [
          "PPT API - Proposal Preparation"
        ],
        "summary": "Create a new proposal",
        "description": "Creates a new proposal in the ODA.",
        "operationId": "create_proposal_ska_oso_services_oso_api_v2_pht_prsls_create_post",
        "security": [
          {
            "SKAO Entra ID Authentication": []
          },
          {
            "SKAO Entra ID Authentication": []
          }
        ],
        "parameters": [
          {
            "name": "X-Request-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Request-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Proposal",
                "examples": {
                  "default": {
                    "summary": "Example Proposal",
                    "description": "A minimal example of a proposal entity.",
                    "value": {
                      "prsl_id": "prp-ska01-202204-02",
                      "status": "draft",
                      "cycle": "5000_2023",
                      "info": {
                        "title": "The Milky Way View",
                        "proposal_type": {
                          "main_type": "standard_proposal",
                          "attributes": [
                            "coordinated_proposal"
                          ]
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "title": "Response Create Proposal Ska Oso Services Oso Api V2 Pht Prsls Create Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ska-oso-services/oso/api/v2/pht/prsls/mine": {
      "get": {
        "tags": [
          "PPT API - Proposal Preparation"
        ],
        "summary": "Get a list of proposals the user can access",
        "description": "List all proposals accessible to the authenticated user.\n\nThe proposals are determined from the underlying data store by:\n1.) Resolving accessible proposal IDs via, list_accessible_proposal_ids,\n2.) Fetching each proposal by ID and \n3.) Returning the proposals as a list (empty if none are found).\n\nReturns:\n    list[Proposal]: Proposals accessible to the current user.",
        "operationId": "get_proposals_for_user_ska_oso_services_oso_api_v2_pht_prsls_mine_get",
        "security": [
          {
            "SKAO Entra ID Authentication": []
          },
          {
            "SKAO Entra ID Authentication": []
          }
        ],
        "parameters": [
          {
            "name": "X-Request-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Request-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Proposal"
                  },
                  "title": "Response Get Proposals For User Ska Oso Services Oso Api V2 Pht Prsls Mine Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ska-oso-services/oso/api/v2/pht/prsls/{prsl_id}": {
      "get": {
        "tags": [
          "PPT API - Proposal Preparation"
        ],
        "summary": "Retrieve an existing proposal",
        "description": "Retrieves the latest proposal by prsl_id.\n\nReturns:\n    Proposal: Returns the latest proposal for the supplied prsl_id, including the metadata.",
        "operationId": "get_proposal_ska_oso_services_oso_api_v2_pht_prsls__prsl_id__get",
        "security": [
          {
            "SKAO Entra ID Authentication": []
          },
          {
            "SKAO Entra ID Authentication": []
          }
        ],
        "parameters": [
          {
            "name": "prsl_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Prsl Id"
            }
          },
          {
            "name": "X-Request-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Request-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Proposal"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "PPT API - Proposal Preparation"
        ],
        "summary": "Update an existing proposal",
        "description": "Updates a proposal in the underlying data store.",
        "operationId": "update_proposal_ska_oso_services_oso_api_v2_pht_prsls__prsl_id__put",
        "security": [
          {
            "SKAO Entra ID Authentication": []
          },
          {
            "SKAO Entra ID Authentication": []
          }
        ],
        "parameters": [
          {
            "name": "prsl_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Prsl Id"
            }
          },
          {
            "name": "X-Request-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Request-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Proposal"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Proposal"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ska-oso-services/oso/api/v2/pht/prsls/batch": {
      "post": {
        "tags": [
          "PPT API - Proposal Preparation"
        ],
        "summary": "Retrieve multiple proposals in batch",
        "description": "AI is creating summary for get_proposals_batch",
        "operationId": "get_proposals_batch_ska_oso_services_oso_api_v2_pht_prsls_batch_post",
        "security": [
          {
            "SKAO Entra ID Authentication": []
          },
          {
            "SKAO Entra ID Authentication": []
          }
        ],
        "parameters": [
          {
            "name": "X-Request-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Request-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_get_proposals_batch_ska_oso_services_oso_api_v2_pht_prsls_batch_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Proposal"
                  },
                  "title": "Response Get Proposals Batch Ska Oso Services Oso Api V2 Pht Prsls Batch Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ska-oso-services/oso/api/v2/pht/prsls/status/{status}": {
      "get": {
        "tags": [
          "PPT API - Proposal Preparation"
        ],
        "summary": "Get a list of proposals by status",
        "description": "Function that requests to GET /proposals/status are mapped to.\n\nRetrieves the Proposals for the given status from the\nunderlying data store, if available\n\nReturns:\n    list[Proposal]",
        "operationId": "get_proposals_by_status_ska_oso_services_oso_api_v2_pht_prsls_status__status__get",
        "security": [
          {
            "SKAO Entra ID Authentication": []
          },
          {
            "SKAO Entra ID Authentication": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Status"
            }
          },
          {
            "name": "X-Request-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Request-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Proposal"
                  },
                  "title": "Response Get Proposals By Status Ska Oso Services Oso Api V2 Pht Prsls Status  Status  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ska-oso-services/oso/api/v2/pht/prsls/reviews/{prsl_id}": {
      "get": {
        "tags": [
          "PPT API - Proposal Preparation"
        ],
        "summary": "Get all reviews for a particular proposal",
        "description": "Function that requests to GET /reviews/{prsl_id} are mapped to.\n\nGet reviews for a given proposal ID from the\nunderlying data store, if available.\n\nReturns:\n    list[PanelReview]",
        "operationId": "get_reviews_for_proposal_ska_oso_services_oso_api_v2_pht_prsls_reviews__prsl_id__get",
        "security": [
          {
            "SKAO Entra ID Authentication": []
          },
          {
            "SKAO Entra ID Authentication": []
          }
        ],
        "parameters": [
          {
            "name": "prsl_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Prsl Id"
            }
          },
          {
            "name": "X-Request-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Request-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PanelReview"
                  },
                  "title": "Response Get Reviews For Proposal Ska Oso Services Oso Api V2 Pht Prsls Reviews  Prsl Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_get_proposals_batch_ska_oso_services_oso_api_v2_pht_prsls_batch_post": {
        "properties": {
          "prsl_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Prsl Ids",
            "description": "List of proposal IDs"
          }
        },
        "type": "object",
        "required": [
          "prsl_ids"
        ],
        "title": "Body_get_proposals_batch_ska_oso_services_oso_api_v2_pht_prsls_batch_post"
      },
...
      "Capabilities": {
        "properties": {
          "mid": {
            "$ref": "#/components/schemas/Mid"
          },
          "low": {
            "$ref": "#/components/schemas/Low"
          }
        },
        "type": "object",
        "required": [
          "mid",
          "low"
        ],
        "title": "Capabilities"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ObservatoryPolicy": {
        "properties": {
          "cycle_number": {
            "type": "integer",
            "title": "Cycle Number"
          },
          "cycle_description": {
            "type": "string",
            "title": "Cycle Description"
          },
          "cycle_information": {
            "$ref": "#/components/schemas/CycleInformation"
          },
          "cycle_policies": {
            "$ref": "#/components/schemas/CyclePolicies"
          },
          "telescope_capabilities": {
            "$ref": "#/components/schemas/TelescopeCapabilities"
          }
        },
        "type": "object",
        "required": [
          "cycle_number",
          "cycle_description",
          "cycle_information",
          "cycle_policies",
          "telescope_capabilities"
        ],
        "title": "ObservatoryPolicy"
      },
...
      "OsdDataModel": {
        "properties": {
          "observatory_policy": {
            "$ref": "#/components/schemas/ObservatoryPolicy"
          },
          "capabilities": {
            "$ref": "#/components/schemas/Capabilities"
          }
        },
        "type": "object",
        "required": [
          "observatory_policy",
          "capabilities"
        ],
        "title": "OsdDataModel"
      },
...
    },
    "securitySchemes": {
      "SKAO Entra ID Authentication": {
        "type": "http",
        "description": "Access token authorization from Microsoft Entra ID.",
        "scheme": "bearer",
        "bearerFormat": "Entra ID Access Token (JWT)"
      }
    }
  }
}